add const iterator version in base matrix class
The current implementation contains only one non-const version of the
iterator. It leads some limitation, for example it is impossible to pass
(compile-time error) const reference on the `matrix_base` instance into
function that doesn't modify it, but operates with the matrix object via
iterators.

It is proposed to add a `const` version of the iterator class.

Test: $ mkdir build
      $ cd build
      $ cmake -DMUNKRESCPP_DEVEL_MODE=ON ..
      $ make full
Issue: #2.
Signed-off-by: Gluttton <gluttton@ukr.net>