overload the validation functions for unsigned types
Values of an unsigned type are valid by default (because they can't be
negative, NaN or infinity).

By default validation is ommited and isn't called by the solver. Also,
probably users won't enable validation for unsigned types explicitly.
But if they will (accidentally or for some reason) this overloading
allows to save several CPU instructions.

Overloading for `is_data_valid` is more important and allows to skip the
whole matrix iteration.

Overloading of the `is_data_invalid` function is less important and
added just for convenience.

Also it is proposed to use `std::signbit` function instead of
comparison with zero.

Test: $ mkdir build
      $ cd build
      $ cmake -DMUNKRESCPP_DEVEL_MODE=ON ..
      $ make full
      $ ./tests/munkrestest
Issue: #4.
Signed-off-by: Gluttton <gluttton@ukr.net>