make adapters interchangeable with their bases
Adapters for OpenCV and Qt matrices use a private inheritance, as a
result they can't be passed into functions that require their base
classes as input arguments.
More formal, the following code won't compile:
    ```
    void use_opencv_matrix (const cv::Mat_ <int> & in);
    munkres_cpp::matrix_opencv <int> m (1, 1);
    use_opencv_matrix (m);
    ```
with the following error: error: ‘cv::Mat_<int>’ is an inaccessible
base of ‘munkres_cpp::matrix_opencv<int>’.

Also for all adapters it is impossible to create a new instance from the
instance of a base class.
More formal, the following code won't compile:
    ```
    cv::Mat_ <int> create_opencv_matrix ();
    munkres_cpp::matrix_opencv <int> m = create_opencv_matrix ();
    ```
with the following error: error: conversion from ‘cv::Mat_<int>’ to
non-scalar type ‘munkres_cpp::matrix_opencv<int>’ requested.

It is proposed to use a public inheritance for all adapters and add
constructors and assignment operators that take an instance of a base
matrix class as in input parameter.

To ensure that issues are fixed and adapters behave in a proper way the
corresponding compile-time tests added.

Test: $ mkdir build
      $ cd build
      $ cmake -DMUNKRESCPP_DEVEL_MODE=ON ..
      $ make tests
Issue: https://gitlab.com/Gluttton/munkres-cpp/-/issues/7
Signed-off-by: Gluttton <gluttton@ukr.net>